+++ src/screens/Welcome/index.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import { View, Text, StyleSheet } from 'react-native';
+
+const Welcome = () => {
+  return (
+    <View style={styles.container}>
+      <Text>Gerencia suas plantas de forma fácil</Text>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+});
+
+export { Welcome };
